Fixed recurring trigger for DBupdate even when there was no change
Summary: Basically, the second list that was created didn't have any value in it's id-name hash set and that led to recurring rename trigger. This revision fixes that by filling the names in the second list as well as applying an empty check on keys.

Test Plan: tested in debugger and well table

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17002