Stopping incorrectly initialized batch programs [devel]
Summary:
The Execute_batch global function is void, thus it is mandatory
to check BP().isOK() after its execution to know whether it succeeded or not.

Test plan:
Ran a batch test program on Linux with or without a valid data root path. Checked
that the instance with an invalid data root does not reach its doWork implementation.