Caching Proj transform object for faster conversion
Summary:
If you export 3D SEGY to another CRS, it is terribly slow because of the repeated construction/destruction (and the underlying sqlite database access) of the transform proj object. Caching it makes the operation much faster. This looks thread-unsafe, but I am in favour of keeping it this way as
1. Adding a lock would add unnecessary overhead
2. Almost all the conversions we do are done during data I/O which are mostly sequential operations.

Test Plan: Exported the default 3D Seismic cube from F3 to SEGY with a different CRS. It takes less than a minute as compared to over an hour earlier.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17016