Fixed missing OpenSSL dependency on Linux [devel]
Modules and plugins that link against the OpenSSL libraries need to know
that these libraries are installed in an 'OpenSSL' subfolder. Otherwise they will
not load.
Does not apply to macOS that seems to handle it in a transparent way.
Does not apply to Windows which does not install the OpenSSL in a subfolder.