Updated directory structure on macOS [devel]
Added a disambiguation of the debug and release targets (libraries and applications),
when using a multi-config generator (Ninja Multi-config, XCode).
The only actual change is for the Debug configuration, where libraries and applications are
placed in a Debug subfolder of the Frameworks and MacOSX folders respectively.
This is a temporary solution to allow have debug and release outputs in a single build directory,
eventually a better solution, for all platforms, would be to add a postfix to the debug targets.

Tested building, installation, packaging and signing of the libraries. od_main starts and the test
programs run.