Added the macro OD_SET_PROG_INSTALL_RPATH [devel]
All GUI apps, batch programs and test programs may need to set the INSTALL_RPATH target property
on Unix.

On APPLE this is required because these applications are in a different folder than the libraries.
This will eventully be the case on Linux as well.

On Linux this is also required for any application that link against OpenSSL, since the OpenSSL libraries
are in a different subfolder.