Moved the loading of the OpenSSL core libraries [devel]
Moved the implementation from Network to General, in the new dedicated class.
Removed an obsolete pErrMsg (caused a forced crash in Debug mode),
fixed the HttpRequestProcess::errmsg_ which was not extracting all the information
from multiple calls to reportError and reportSSLErrors (received from Qt).

Added a test to connect with https to a site which does not have a valid certificate.
We should get an API to optionally bypass such errors, like the flag --no-check-certificate of the
command wget.