Adjusted implementation of File::getTimeInSeconds
Summary:
QFileInfo::birthTime does not return anything good on Unix, but
QFileInfo::metadataChangeTime() returns the same creation date as seen on the Linux
file browser dolphin, for files.
For folders it returns the same as lastmodif, which is not perfect, but still better than
an undefined date.

Test Plan: tested on Linux and Windows by comparing with the file explorer on files and folders.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D17097