now updating the transparency to non-transparent
Summary:
Earlier the default was no color ( black with full transparency ). Now while displaying if the color is flagged as no color, only the transparency value of the local copy of the data is changed
OD-5466

Test Plan:
Tested, now displaying the data in native color.
Also tested by changing the background to blue and black and deleted the display file. The track color was contrasting color to the scene, for Black scene it was red, for Blue scene it was dark shade of black

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd, raman

Differential Revision: https://phabricator.dgbes.com/D17102