Adjusted implementation of File::getTimeInSeconds
Summary

QFileInfo::birthTime does not return anything good on Unix, but
QFileInfo::metadataChangeTime() returns the same creation date as seen on the Linux file browser dolphin, for files.
For folders it returns the same as lastmodif, which is not perfect, but still better than an undefined date.

Applied the same changes to LocalFileSystemAccess::timeCreated and LocalFileSystemAccess::timeLastModified,
setting the return value to -1, which is a more clearer undefined value.
Using a much simplier implementation for Time::getTimeInMilliSeconds(), based on QTime::msecsSinceStartOfDay()

Test Plan

tested on Linux and Windows by comparing with the file explorer on files and folders.