Small fix in OD::PythonAccess::GetPythonEnvPath
In the absence of store python settings, the function OD::PythonAccess::hasInternalEnvironment
should be used to determine if the default source is internal or system.

This fixes an incorrect initialization of the python toolbar in the absence of user settings,
when using the python bundle installed by the OpendTect installer.

[Jira: OD-4728]
[Jira: OD-5483]