Fixed updated errors from the Python settings window
Summary

Two distinct issues, platform independent, but much more visible on Windows because the python commands take a much longer time to run.

1- Race issue in OD::PythonAccess::verifyEnvironmentCB: The class member moduleinfos_ was changed after releasing the lock, for instance if the user goes to the python settings and pressed the test button soon after having started the application. In that case the initial background tests are still running, hence one must make a copy of moduleinfos_ before releasing the lock, and then stop using the class member functions in verifyEnvironmentCB.
2- The python settings dialog was in most case no longer committing the changes within the acceptOK function, because the call to OD::PythA().isUsable was most often not hit. This is coming from the chgdsetts_ class member, which has become unreliable: Multiple consecutive calls to parChgCB could result in emptying this IOPar, make the system believe that no change had occurred. Solving it by simplifying the structure: To check if settings have changed, exclusively comparing the IOPar read from the screen GUI with the one saved as initialsetts_.

Also fixed an incorrect initialization of the python toolbar in the absence of user settings, when using the python bundle installed by the OpendTect installer. In the absence of store python settings, the function OD::PythonAccess::hasInternalEnvironment
should be used to determine if the default source is internal or system.

[Jira: OD-5483]

Test Plan

Starting od_main, going to the python settings dialog, hitting the test button, closing (OK or Cancel), going to the Machine Learning control center -> Should redo the tests if the settings have changed, and the status bar text must match the Python settings entered in the corresponding dialog.
Also checked the workflow from OD-4728, to ensure no regression is made.