Updated cmake support of the OpenSSL libraries on Linux [devel]
Summary

We need a runtime support of the OpendTect libraries that are dependent on OpenSSL.
That means that there must be a repository telling us if an OpendTect library is dependent on OpenSSL,
because it is not possible to open that library with the SharedLibAccess to read that information from
the PluginInfo external function: The library will not load on Linux (its SharedLibAccess will not be OK),
unless the OpenSSL libraries (libcrypto and libssl) are loaded beforehand. At this loading is very much
dependent on the Linux flavor: It should only be done if the system does not have libraries of the same version
as the OpenSSL version used to build the OpendTect libraries, which means, at the time of writing:
1- CentOS 7, Rocky 9 and Ubuntu 22.04 will load the OpenSSL libraries provided with the OpendTect installation,
as they are using the system OpenSSL v1.0.2, 3.0.7 and 3.0.2 respectively.
2- Rocky 8, Ubuntu 20.04 and OpenSuse 15.5 will use the system libraries v1.1.1

Test Plan

Works on OpenSuse, to be tested with a release on the other Linux flavors.