Some optimizations for Scene::objectMoved
Summary:
If you have a large number of objects (say 2D horizons) in the scene, 'show all' and 'hide all' operations were too slow. The reason was the inefficient objectMoved call.
1. Objects that do not have a movement Notifier should not emit objectMoved.
2. The receiving object should get a valid moved object ID to decide what action to take.

Test Plan: 2D horizons are faster to load, show, hide. Also loaded a bunch of 3D horizons and Faults on sections only and Moved the Inline/Xline sections. The intersections are updated correctly and smoothly.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17162