Fixed loading of some plugin classes [devel]
Even if the function of one plugin cannot be used, loading of that function in the other plugins should still be done, as the plugins are supposed to be independent.
This should fix the absence of the SEG-Y and well survey info providers in some installations, where the dGBOpenVDS library does not load successfully.