Added FilePath::localPath [devel]
Summary:
added a function to output the path of a FilePath, but without
prefix, domain and postfix it is a URI
Also fixed the insert function which would wipe out the prefix, domain and postfix, and a minor other bug fix in pathOnly, that did not correctly use the Style argument.

Test Plan: expanded the test program, which presents examples of the various FilePath function outputs.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D17220