Fix OD5549 Crash uiWellLogDispProperties::doPutToScreen()
Summary:
Fixes OD-5549 [CR 20240527110824] uiWellLogDispProperties::doPutToScreen()
Not able to reproduce crash on  Linux but saw repeatable crash on Windows when pressing Save button
in uiWellDispPropDlg. The code keeps a pointer to the Well::Data display parameters object which becomes
invalid after display paramaters are reloaded.
Now using the Well::Data notifier to ensure the pointer to the display parameters is updated after a reload.

Test Plan:
Tested on Linux:
- no crash observed on Windows after pressing Save button
- UI on both Windows and Linux seems to be working ok

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17227