Minor rework in file/filesystemaccess [devel]
Summary:
Enabled copy operations between FileSystemAccess implementations.
Will only work if either from or to protocols support it.
BufferString -> uiString conversion
Added a TaskRunner pointer to the File copy and copyDir functions, to be
forwardded to the FileSystemAccess implementation if it can make use of it.

Changed the default value of four virtual functions of the FileSystemAccess
base class to false. Should only need to be overriden by the implementation when
they add support for it.

[Jira: OD-5571]

Test Plan: No real change in the base functionality. Compiles and links.

Reviewers: raman, arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D17228