Porting from main: Fixes for uiMnemDefLogDialog.
Summary:
Setting default log for a Mnemonic to "none" did nothing, when it should remove any default associated with the mnemonic.

Fixed a crash in MnemDefLog dialog

If the log is set to "None", in that case the changedlog_ is then not modified and remains nullptr. The nullptr check then handles the crash.

Test Plan:
tested in Well Table

tested in well table

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com/D17232