Fixed all the crashes related to Custom Mnemonics
Summary:
Jira issue: OD-5568

This revision fixes this Jira issue as well as some more crashes related to Custom Mnemonics. The issue primarily was that Logs with such mnemonics assigned to them would not be aware of the changes made in the mnemonics. This is further exacerbated by the fact that existing mnemonics are deleted and then rewritten always when using the custom mnemonics dialog. This will need fixing.

Hence, MnemonicSet class has been made a CallBacker and now emits a notifier for mnemonicChanged. This is caught by Logs and then further passed on to LogSet to update default mnemonics dataset as well.

Tested, works

Test Plan: tested in well table

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com/D17280