Restructured the odbind python package [devel]
odbind was not available in a standard development environment.
Deploying the same structure as for odpy and dgbpy, which means:
1- Forwardding the path to the python source files from the CMakeLists.txt file as a compile definition
2- Using the forwardded path to call OD::PythonAccess::addBasePath when necessary
3- Loading the C++ plugin for the applications that need to be able to access (import) the odbind python plugin

This can only work if the path the contains the odbind package contains a folder by that name, meaning that this folder
needed to be created and the python files moved inside it.
This subdirectory structure is also used by other projects like OpenVDS and OpenZGY.
Also simplified the installation of the python package.