Silently adding required property synthetics in StratSynth::DataMgr
When adding a synthetic definition which is based on a filtered property, tentatively
adding now a definition for its input, providing that the property reference can be found
in either the property list of the current layer model, or the overall property reference list.

SynthGenParams::getRef will now also return a valid property reference for the synthetic definitions
of type FilteredStratProp, based on the input name.