Fixed incorrect filepath in json objects
Summary:
file paths set in json objects and read from json objects should always be using
Unix file delimiters, otherwise they won't be parsed correctly for instance inside python.
The current logic did not take into account the URI/URL types of file paths, as a result a file path retrieved from a json object would not be recognized as a URI/URL type of file path on Windows.
It is not possible to use FilePath::fullPath( FilePath::Unix ), as this function always appends '/cygdrive/' at the start of the filepath as part of the cleanup, and does not do much without cleanup, as FilePath::dirUpTo does not use the style, hence returns a string in local style. Maybe we should consider removing it.

[Jira: OD-5648]

Test Plan: in progress, download of pre-trained models work now.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17302