Fixed distributed processing issue under a proxy
Summary:
The application proxy was not set for applications that did not use the Network module
for downloading a file, like the OpendTect installer.
It should be done for any application that uses either a TCP server or socket. Also the ProxyType 'HttpProxy'
does not allow listening and outgoing connections, which is required, thus switching to to the type 'DefaultProxy'.
This effectively solves the distributed computing error, where the processing nodes fail to open a listen socket.

Test Plan:
ran the install target, and used either od_SeisMMBatch, or the test programs test_listening_server
and test_client_connection for debugging. The proxy is installed at the system level, and the distributed computing
works with and without entering the proxy settings in the OpendTect GUI (thus having proxy setting stored or missing
from the user settings).
In both settings, the OpendTect installation manager is capable of downloading the definition files.

[Jira: OD-5647]