Fix for Horizon intersection displayed on 'turned-off' sections
Summary: After the optimization introduced in https://github.com/OpendTect/OpendTect/commit/45416e350dc9efafc1b396313b8d0632bce873ef, the objectMoved function and hence the updateIntersectionLines functions receives a valid VisID of the object that has moved. In case of a section being turned off, this ID is valid but the corresponding object is not in the list of active objects. The function now takes care of this case by simply removing the intersection lines on that section.

Test Plan: Tested by displaying a 3D horizon on a bunch of sections in section only mode. Then turned off/on the sections and moved some of them. The intersection updates with every move and turn on/off. 2D horizons were not affected by this bug.

Reviewers: arnaud, nanne

Differential Revision: https://phabricator.dgbes.com/D17307