Sorting ListBox items now preserves all its properties(Pixmap, color etc)
Summary:
Sorting the listbox would, erase the original items and repopulate it with new items using sorted itemnames. This meant that only the name of original item ws preserved or its marked status, as that was being cached before sorting, and the rest of its propeties like Pixmap or background color would not be preserved.

this revision fixes that. sorting is now handled in the listboxbody class and the original items are preserved.

Test Plan: tested in a listbox in Stratigraphy manager

Reviewers: nanne

Reviewed By: nanne

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17316