Fix OD5695 Dist Proc Setup dialog only allows subnets of 255.255.255.0
Summary:
Fixes OD-5695 Distributed Processing Setup dialog only allows subnets of 255.255.255.0
Added a field to specify the subnet prefix length in the Advanced Settings dialog
of the Distributed Processing Setup. A value of 0 means a mask of 0.0.0.0 (all), 8 means a
mask of 255.0.0.0, 16 means a mask of 255.255.0.0 and 24 means a mask of 255.255.255.0.
Maybe it would be more intuitive to have a widget to actually enter the mask but I couldn't
see any existing widget that could easily be adapted for that role. -- Maybe a listbox would be ok since there are only 4 values.

Was able to interact with a host at 172.18.0.2 from od_BatchHosts on 192.168.1.107 using prefix length
of 0. Still need to test if multi machine procesing is possible with these disparate IP addresses.
Will test that next week.

Test Plan:
Tested on Linux:
- was able to interact with a host on 172.18.0.2 from Distributed Processing Setup dialog
running on a host on 192.168.1.107 by setting the subnet prefix of 0
- verified setting saved to BatchHosts file
- verified setting restored from BatchHosts file when starting od_BatchHosts

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17325