Fix for Horizon display on sections with no attributes displayed
Summary: Consistent with the earlier behaviour in OD, the sections (2D and 3D) with no attributes enabled are deemed turned off and therefore no Horizons (2D and 3D) displayed in 'sections-only' mode should be displayed on them.

Test Plan: Tested with both 2D and 3D horizons and sections. When multiple attributes are displayed on a section, turning off all the attributes turns off the intersction lines as well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17333