Fix for partial processing of 2D T2D
Summary:
When the velocity cube is not pre-loaded, the velocity traces coming from the reader where not checked against the TrcKey given by the iterator. Now, the iterator is not used in this case. A new SeisTrcReader is used with the SelData correponding to the chunk and the TrcKey is taken from the incoming traces, thereby ruling out the possibility of a mismatch.
Fixes OD-5696.

Test Plan: Tested 2D and 3D time-depth conversion, in both on-the-fly and batch mode.

Reviewers: nanne

Maniphest Tasks: T2

Differential Revision: https://phabricator.dgbes.com/D17344