Porting from main: uiMarkerList class now declared in header file
Summary:
Just as the title mentions, uiMarkerList is now declared in header file so that it can be used in other places.

This is going to be used in Advance Well Table. Hence is a part of:
Jira Issue OD-5588

Made it compile

Test Plan: tested in Basic Well Manager as well as Advanced Well Manager.

Reviewers: raman

Reviewed By: raman

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com/D17336