Put offset in object name, not just the index
Summary:
The treeitem name showed something like 'offset index #', where # is the
index of the offset slice. The component name is 'Offset: #' where # is the
actual offset. Using the component name for the userref now.

Also fixed a crash when the number of attribute outputs differs from the
number of components in the trace.

[Jira: OD-5712]

Test Plan:
No crash anymore when loading Penobscot prestack data. Treeitem shows new name,
with offset.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com/D17354