Added an error message when configuring the Qt dependencies [devel]
If Qt::Core is found, but some package component was missing, nothing
was reported during the configuration of the project.
This added message should warn the user that the Qt installation may not
be completely right, or that it needs to be complemented by additional
packages.

This is for instance the case on macOS with Qt6, where the
Qt::PrintSupport target is dependent on cups.