Preparation for the release of Universal binaries on macOS [devel]
Summary:
Merging back the macarm and macintel plfsubdirs into the former mac plfsubdir.
Releases will exclusively produces universal binaries, containing both arm64 and intel
binaries.
As a result the enum in OD::Platform::Type is back to only three types, made an enum class out of it.
However, plfdefs.h keeps setting __macarm__ OR __macintel__ on macOS, in case they are needed in any implementation. For instance, I want the about dialog to contextually show which type of binaries are being executed.

Test Plan: compiles on Linux, will need testing on macOS when it is landed.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D17395