Moved the (Qt) PrintSupport to a dedicated plugin [devel]
Summary:
Removed the dependency to Qt::PrintSupport of uiBase. Most of the simple GUI applications do not need it, like the log viewer, the installer, ...
For the applications that need it, like od_main, using a plugin avoids a direct dependency to the corresponding Qt library, and all of its dependencies. In this way we will be able to uncouple for instance the OpendTect installer from QtPrintSupport and its dependencies like cups on macOS.
Added a cmake variable that would enable building without this support.
Not using the ALO system to load the plugin, that is taken care of by the initialization of the uiBase base module.

Test Plan: compiles, links, saving to PDF works as previously.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17398