Moved the QSql database implementations to a plugin [devel]
Summary:
Removed the Database base module:
Moved the non-Qt related base classes and factories in General,
moved the implementation in the new ODQSql plugin.
There should be not speed difference between loading a base module (Database)
or a plugin (ODQSql), as in both cases it is one additional runtime library to load. The size of the ODQSql plugin should be smaller than the former Database library.
Main advantages are thus:
1- No dependency of any application to the Qt::Sql library (and its eventual depencies), which is especially useful for the batch programs (for now) and the basic GUI applications, including the OpendTect installer.
2- All applications but od_main won't load the plugin. Again using the program context for that, not the ALO system.

Test Plan: compiles and links with or without enabling the usage of Qt::QSql

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17403