fixed warning pixmap disappeared if list box was emptied and refilled
Summary: in seismic manager any call to fillListBox function in uiIOObjSelGrp empties the list box and refill it. This means that the warning pixmap shown for objects that were not OK, disappeared. This revison attaches a callbacker to listChanged notifier in uiIOObjSelGrp and calls the fucntion to check if all the objects are OK and set the pixmap again.

Test Plan: tested in Seismic manager by clicking on all the menu buttons and seeing that pixmaps are restored. Also, tested with 73 Seismic cubes to see if the operation is too costly.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17427