Removed the uiGoogleIO plugin [devel]
Summary:
Moved the corresponding classes to uiIo, uiSeis and uiWell, to make them accessible to
modules non dependent on uiODMain.
Also reworked the existing classes, they were too many classes doing almost the same work.
Most of the dialogs were designed to export a single object at the time, but this is wrong,
they should handle a set of objects of the same type, and be accessed from the parent tree items,
not the individual tree items.

Also fixed an issue with the uiListBox usage for the 2D line selection: If one wants at least one line to be
selected, 'setChosen' was not working because the listbox was not able of deselecting all items.

Test Plan:
Exported 2D lines, well positions, random lines, poinsets, polygons, in KML and GeoJSON formats.
Mostly good, still some remaining issues that already existed before with the GeoJSON output.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17428