Use MultiWellReader to read well data first before loading in 3D scene
Summary:
1) using MultiWellReader to get all the well data first and going through it to remove potential wells that won't load.
2) Removed reloading of wells for loading into 3D scene. This is because usually the one in memory is the latest iteration of the well, not the one on the disk

This has helped in accumulating all the error messages for potential failures and only show them at the end of load operation without interrupting and stopping loading of the wells.

Test Plan: tested by loading wells in 3D scene before Well::MGR() is initialized as well as after the manager is initialized.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17437