Removed the checksum (crc64) function [devel]
Summary:
First the test_checksum function is flagged by the Bitdefender antivirus (False positive)
However the 12 years-old implementation seems outdated. New projects, like the following two,
return different results for the input test string "123456789":
https://toolkitbay.com/tkb/tool/CRC-64
https://github.com/lordmulder/CRC-64/tree/master

This suggests a bug in the old implementation. However if with that bug removed, the crc64 checksum
is still not guaranteed to be unique, unlike the cryptographic hashes which are now available.
Thus switching the implementation of System::macAddressHash to a sha3-256 sum. The output is a string,
not an integer.

Test Plan: compiled, test_systeminfo is OK.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17438