Minor fix in the packages creation for Unix [devel]
The default behaviour of 7z for storing symbolic links as links seem to
have changed between v17 and v24. It was enabled by default in v16/17,
by disabled by default in v24.
We always want this option to be enabled, thus adding it to the
packaging command line.

Will reduce the package size on both Linux and macOS.