Added verbosity when missing some Qt components during cmake configuration
Repeated a find_package call with the 'REQUIRED' keyword to get the full
verbosity of what is missing when it fails.
The only disadvantage of that solution is that it stops the
configuration, instead of collecting all messages and printing them all
at once at the end, but for Qt we do not have much choice.