Fixed several issues with the single horizon 2D import/export
Summary

For the import, for single 2D horizons, all horizons loaded in a session were having the data
of the first imported horizon, because the scanner_ class member was not deleted on file change.
For the export it was impossible to change the input horizon selection: uiIOObjSel::fillEntries()
was calling uiIOSelect::setEntries, and its workctio_ was good, but the uiIOSelect::inp_ was reset
by the call to uiIOSelect::setEntries. Obvisouly this function should remember the previous selection,
and restore it if possible.
Also fixed several memory issues in the handling of the ObjectSet class, and HiddenParam usage.

Added uiFileInput::defaultExtension()
Setting "txt" as the default extension for all uiASCIIFileInput
Moved several Setup ctor/dtor to the cc files
Various code style updates

For Horizon 3D, Horizon 2D, Fault, FaultStickSet, FaultSet, bulk and single object exports:
Fixed the default output unit of measure: Should be the user UoM, not the storage UoM, thus "ms" and not "s"

For their single object export only:
Setting the output filename based on the to-be-exported dataset name (only the file name, path does not change)

[Jira: OD-5848]

Test Plan

Can now export single 2D horizons, and import them.