Better memory information in crash reports
Summary: The system memory information in carsh reports is fetched by the crash reporter application, and by the time this is done, the main OpendTect application is already gone and the memory used by it is also freed up almost entirely. Now, the crashed program itself sends free memory information to the crash reporter as a command line argument.

Test Plan: Preloaded a large volume and forced a crash. The crash reporter window shows correct memory information before and after the crash.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D17454