Large rework of the MPE::Engine [devel]
Summary:
Handled the following issues:
1- The MPE::Engine was still relying on the datapacks being present in the global datapack
manager, instead of pulling them from the vizualization objects
2- The MPE::Engine was partially owning (ref/unRef) the trackers and editors, whereas it should only observe them
3- The MPE::Engine get/set functions for the trackers and editors were entirely index based.
This structure cannot be kept when using a WeakPtrset which has a dynamic size, it is resized each time an object gets
deleted. Eventually there was no real need of using an index based structure.
4- The EM::ObjectID is the only identifier to be used for finding an MPE::EMTracker or MPE::ObjectEditor in the system,
the visualization object should never try to create them before getting a valid EM::ObjectID assigned.
5- The MPE::EMTracker objects, unlike the MPE::ObjectEditor objects, where not stored by the vizualization items. This is not
correct, the vizualization objects (from visSurvey and/or uiViewer2D) are the main and only objects that should create, own and
remove them. This allows removing entirely the ref/unRef tracking in the MPE, uiVis and uiODMain modules for these objects.
6- Reworked the factories: There should not be one global factory for all EM::Object, by factories by data type. It does not
make sense to have in a single factory a 3D horizon editor and a polygon-body editor.
The same applies for the GUI factories, the uiSetupGroupFactory was not even using the base factory class.
7- The random lines was not yet avoidding to use the global datapack manager, made them use the same implementations as
3D slices, 2D lines and 3D volumes. Made some virtual functions more generic, to accept a SeisDataPack object instead of
a RegularSeisDataPack, as the first will allow working with a RandomSeisDataPack as well.

Test Plan:
Tracking of 3D horizons in the 3D scene on inlines, random lines and Z slices OK. Resume tracking as well.
Tracking of 2D horizons in the 3D scene OK
Picking of fault, faultstickets and polygon-based bodies in the 3D scene OK
Tracking of horizons in the 2D viewers OK.
Picking of Faults and Fault Stick Sets in 2D viewers has issues.
Session restore to be tested

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17460