Fixed: isLoaded function of Well::Info always returned true
Summary:
The order of conditions was such that if track would be read before D2T, then the getInfo function call can't be reached. This caused info to not be read, but replvel_ to be set.
This is very wrong because the get function in Well::Manager won't ever load the info because it excludes reqs for componenets that have already been loaded. Having a replvel_ that is different from the default velocity will lead to a true return on Well::Info::isLoaded call.

This caused issues in Basemap, where bounding box would start from (0,0) coordinate and put all the wells there, as surface coordinate( more importantly whole info section ) was not read.

Test Plan:
1) tested in debugger by checking function calls and values.
2) All the wells are at correct location in the Basemap, when View All is pressed, as opposed to being yanked to (0,0) coordinate like before.

Reviewers: arnaud, raman

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17482