Merged VolumeDataPack with SeisDataPack [devel]
Summary:
The former VolumeDataPack was a single component version of SeisDataPack,
besides VolumeDataPack and SeisDataPack were unrelated, and SeisDataPack in General
was not a good name.
Porting the merging of these two base classes from od6.7
Influence on higher classes is small, as the really important classes are RegularSeisDataPack
and RandomSeisDataPack.
Added three typedef to make legacy code compile without the need of any code change.

Test Plan: displayed seismic data in 2D and 3D viewers, with and without pre-loading

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17490