Added OpendTect_INST_DIR at the project top level
The cmake variable OpendTect_INST_DIR is now set at the root project level,
and exported to FindOpendTect.cmake, such that external plugin builders
can read that variable and set the installation directory in packages
accordingly.
It represents the folder name at the top of the file tree for all
installation packages.