Fixed an issue about well name in Well copier class for HDF5 format
Summary:
Well name is saved in the new HDF5 format in the info section, whereas in the legacy format, it was read from the file name. So, for copied wells, it is important that we set the name of new Well Data object where data is copied to.

Porting from main

Test Plan: tested in Well Table.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com/D17507