bulk change to OMF file for removal of entries
Summary:
opendtect should have the ability to delete objects in bulk. Currently this is done by individually altering omf files for each object. This means locking, reading, editing and writing steps for each object individually.
This revision adds functions to IOMan and IODir class which take a TypeSet<MultiID>. This typeset is then used to remove entries from omf file all at once. Then a new CNotifier has been added called entriesRemoved which takes, again, a TypeSet<MultiID> as its payload. This way those listening on this notifier also have the idea of the total scope of work they have to carry out to be up to date instead of just waiting on the Notifier for next entry and taking relevant actions individually.

The immediate result of this, as seen in wells deletions, is a significant speed improvement in just the delete operation itself not to mention sunsequent clarity and improvements in actions that could be taken later on in other related components.

https://dgbes.atlassian.net/browse/OD-5936

Test Plan:
Tested by deleting 1022 wells. Old deleting time was 2 minutes 30 seconds, With this change, it happens in around 10 seconds.
Also added breakpoint to doWrite function in IODir class (writes omf files). Earlier this was being called for each object to be removed. Now it is called at the end, once and removes all the entries in one go.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com/D17508